+++ src/test/java/cart/integration/CartIntegartionTest.java
@@ -1,30 +1,15 @@
-package cart.controller;
+package cart.integration;
 
 import cart.controller.dto.AddCartRequest;
-import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import org.apache.http.HttpStatus;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.test.context.jdbc.Sql;
 
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.Matchers.hasSize;
 
-@Sql(value = "/test.sql")
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class CartControllerTest {
-
-    @LocalServerPort
-    private int port;
-
-    @BeforeEach
-    void setUp() {
-        RestAssured.port = port;
-    }
+class CartIntegartionTest extends IntegrationTest {
 
     @DisplayName("GET /cart/products 요청시 존재하는 아이디 비밀번호일 시 OK 반환")
     @Test
