+++ src/test/java/cart/integration/ViewIntegrationTest.java
@@ -1,32 +1,15 @@
-package cart.controller;
+package cart.integration;
 
-import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import org.apache.http.HttpStatus;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.test.context.jdbc.Sql;
-
-import java.net.URISyntaxException;
 
 import static io.restassured.RestAssured.given;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-@Sql("/test.sql")
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class ViewControllerTest {
-
-    @LocalServerPort
-    private int port;
-
-    @BeforeEach
-    void setUp() throws URISyntaxException {
-        RestAssured.port = port;
-    }
+class ViewIntegrationTest extends IntegrationTest {
 
     @DisplayName("GET / 요청 시 Status OK 및 HTML 반환")
     @Test
